mcp/requirements.txt
# CobbleAI MCP Tool Server
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==46.0.3
et_xmlfile==2.0.0
fpdf2==2.8.7
geopandas==1.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mcp==1.26.0
numpy==2.3.0
O365==2.0.37
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.3
sniffio==1.3.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
uvicorn==0.41.0
